backend/requirements.txt
alembic==1.5.4
beautifulsoup4==4.6.3
bleach==3.3.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Deprecated==1.2.11
elasticsearch==7.10.1
Flask==1.1.2
fusionauth-client==1.23.0
html5lib==1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
mysqlclient==2.0.3
oauthlib==3.1.0
pkce==1.0.2
PyYAML==5.3.1
requests==2.21.0
requests-oauthlib==1.3.0
translitcodec==0.7.0
urllib3==1.24.3
Werkzeug==1.0.1
wrapt==1.12.1
